permitta/src/repositories/test/data_object_repository_test.py
from database import Database

from ..src.data_object_repository import DataObjectRepository


def test_get_all_unique_attributes(database: Database) -> None:
    repo: DataObjectRepository = DataObjectRepository()

    with database.Session.begin() as session:
        attributes = repo.get_all_unique_attributes(session=session)
        assert len(attributes) == 12

        # check they are all unique
        unique_key_values: list[str] = []
        for attribute in attributes:
            unique_key_values.append(
                f"{attribute.attribute_key}={attribute.attribute_value}"
            )
        assert len(set(unique_key_values)) == len(attributes)

        # with a search term on the key
        attributes = repo.get_all_unique_attributes(session=session, search_term="Sal")
        assert len(attributes) == 3

        # with a search term on the value
        attributes = repo.get_all_unique_attributes(
            session=session, search_term="Restricted"
        )
        assert len(attributes) == 4
